Updated README.md
Made changes to make it a little prettier, though that is subjective… 😬
Made other changes about orthography which are not subjective, being RAE (https://dle.rae.es) the source of my corrections.
I tried not to touch the essence of the repo, but well, I may have gone too far in some modifications without noticing. So, whatever change you (Joaco) may want to revert, I'll revert it. You're the boss.